lib/codegen: ported from jit to mcjit.
jit has been dropped as of LLVM 3.6 (deprecated in 3.5) in favor of the fancier
MCJit (read http://llvm.org/docs/MCJITDesignAndImplementation.html). To make it
work I've also had to make the global Module a unique_ptr, which is for the
best anyways.

Since I haven't tested this on 3.5, I'm increasing the LLVM minimum version
up to 3.6, since I'm not gonna be shy on my own experimental projects :)